+++ lib/decider.rb
@@ -4,82 +4,95 @@ module Decider
   StateAlreadyDefined = Class.new(StandardError)
   StateNotDefined = Class.new(StandardError)
 
-  class Composition < Array
-    BLANK = Object.new
-    private_constant :BLANK
+  Pair = Data.define(:left, :right)
 
-    def initialize(left, right)
-      super([left, right]).freeze
+  class Left
+    attr_reader :value
+
+    def initialize(value)
+      @value = value
     end
 
-    def with(left: BLANK, right: BLANK)
-      Composition.new(
-        (left == BLANK) ? self.left : left,
-        (right == BLANK) ? self.right : right
-      )
+    def deconstruct
+      [:left, value]
     end
 
-    def left
-      self[0]
+    def ==(other)
+      value == other.value
+    end
+  end
+
+  class Right
+    attr_reader :value
+
+    def initialize(value)
+      @value = value
     end
 
-    def right
-      self[1]
+    def deconstruct
+      [:right, value]
+    end
+
+    def ==(other)
+      value == other.value
     end
   end
 
   class Module < ::Module
+    DECIDE_FALLBACK = proc { [nil, proc { [] }] }
+    EVOLVE_FALLBACK = proc { [nil, proc { state }] }
+
+    Decide = Data.define(:command, :state)
+    Evolve = Data.define(:state, :event)
+    Terminal = Data.define(:state)
+
     def initialize(initial_state:, deciders:, evolutions:, terminal:)
       define_method(:initial_state) do
         initial_state
       end
 
-      define_method(:commands) do
-        deciders.keys
-      end
-
-      define_method(:events) do
-        evolutions.keys
-      end
-
-      define_method(:decide!) do |command, state|
-        case deciders.find { |key, _| key === command }
-        in [_, handler]
-          handler.call(command, state)
-        else
-          raise ArgumentError, "Unknown command: #{command.inspect}"
-        end
-      end
-
       define_method(:decide) do |command, state|
-        case deciders.find { |key, _| key === command }
-        in [_, handler]
-          handler.call(command, state)
-        else
-          []
-        end
-      end
+        context = Decide.new(command: command, state: state)
 
-      define_method(:evolve!) do |state, event|
-        case evolutions.find { |key, _| key === event }
-        in [_, handler]
-          handler.call(state, event)
-        else
-          raise ArgumentError, "Unknown event: #{event.inspect}"
-        end
+        deciders.find(DECIDE_FALLBACK) do |args, _|
+          case args
+          in [Proc => fn]
+            context.instance_exec(&fn)
+          in [ctype]
+            command in ^ctype
+          in [ctype, stype]
+            [command, state] in [^ctype, ^stype]
+          else
+            false
+          end
+        end => [_, handler]
+
+        context.instance_exec(&handler)
       end
 
       define_method(:evolve) do |state, event|
-        case evolutions.find { |key, _| key === event }
-        in [_, handler]
-          handler.call(state, event)
-        else
-          state
-        end
+        context = Evolve.new(state: state, event: event)
+
+        evolutions.find(EVOLVE_FALLBACK) do |args, _|
+          case args
+          in [Proc => fn]
+            context.instance_exec(&fn)
+          in [etype]
+            event in ^etype
+          in [stype, etype]
+            [state, event] in [^stype, ^etype]
+          else
+            false
+          end
+        end => [_, handler]
+
+        context.instance_exec(&handler)
       end
 
       define_method(:terminal?) do |state|
-        terminal.call(state)
+        context = Terminal.new(state: state)
+
+        context.instance_exec(&terminal)
       end
     end
   end
@@ -93,7 +106,7 @@ module Decider
       @initial_state = DEFAULT
       @deciders = {}
       @evolutions = {}
-      @terminal = ->(_state) { false }
+      @terminal = proc { false }
     end
 
     def build(&block)
@@ -125,12 +138,12 @@ module Decider
       @initial_state = state
     end
 
-    def decide(command, &block)
-      deciders[command] = block
+    def decide(*args, &block)
+      deciders[args] = block
     end
 
-    def evolve(event, &block)
-      evolutions[event] = block
+    def evolve(*args, &block)
+      evolutions[args] = block
     end
 
     def terminal?(&block)
@@ -146,37 +159,32 @@ module Decider
 
   def self.compose(left, right)
     define do
-      initial_state Composition.new(left.initial_state, right.initial_state)
+      initial_state Pair.new(
+        left: left.initial_state,
+        right: right.initial_state
+      )
 
-      left.commands.each do |klass|
-        decide klass do |command, state|
-          left.decide(command, state.left)
-        end
+      decide proc { [command] in [[:left, _]] } do
+        left.decide(command.value, state.left).map { Left.new(_1) }
       end
 
-      right.commands.each do |klass|
-        decide klass do |command, state|
-          right.decide(command, state.right)
-        end
+      decide proc { [command] in [[:right, _]] } do
+        right.decide(command.value, state.right).map { Right.new(_1) }
       end
 
-      left.events.each do |klass|
-        evolve klass do |state, event|
-          state.with(
-            left: left.evolve(state.left, event)
-          )
-        end
+      evolve proc { [event] in [[:left, _]] } do
+        state.with(
+          left: left.evolve(state.left, event.value)
+        )
       end
 
-      right.events.each do |klass|
-        evolve klass do |state, event|
-          state.with(
-            right: right.evolve(state.right, event)
-          )
-        end
+      evolve proc { [event] in [[:right, _]] } do
+        state.with(
+          right: right.evolve(state.right, event.value)
+        )
       end
 
-      terminal? do |state|
+      terminal? do
         left.terminal?(state.left) && right.terminal?(state.right)
       end
     end